Renamed FloatValueType.cs to DecimalValueType.cs for consistency.

+++ OffByOne/src/OffByOne.Qls/CustomQlsVisitor.cs
@@ -215,7 +215,7 @@
 
         public override AstNode VisitFloatType(QlsGrammarParser.FloatTypeContext context)
         {
-            return TypeConstants.FloatType;
+            return TypeConstants.DecimalType;
         }
 
         public override AstNode VisitMoneyType(QlsGrammarParser.MoneyTypeContext context)
